requirements.txt
flask
python-dotenv
google-generativeai
requests
PyPDF2
python-docx
